Refactor payments form items into separate component
