+++ packages/venia-concept/src/components/Checkout/paymentsFormItems.js
@@ -1,0 +1,178 @@
+import React, { useCallback, Fragment } from 'react';
+import { useFormState } from 'informed';
+import { array, bool, func, shape, string } from 'prop-types';
+
+import BraintreeDropin from './braintreeDropin';
+import Button from 'src/components/Button';
+import Checkbox from 'src/components/Checkbox';
+import Field from 'src/components/Field';
+import TextInput from 'src/components/TextInput';
+import {
+    isRequired,
+    hasLengthExactly,
+    validateRegionCode
+} from 'src/util/formValidators';
+import combine from 'src/util/combineValidators';
+
+/**
+ * This component is meant to be nested within an `informed` form. It utilizes
+ * form state to do conditional rendering and submission.
+ */
+const PaymentsFormItems = props => {
+    const {
+        cancel,
+        classes,
+        countries,
+        isSubmitting,
+        setIsSubmitting,
+        submit,
+        submitting
+    } = props;
+
+    // Currently form state toggles dirty from false to true because of how
+    // informed is implemented. This effectively causes this child components
+    // to re-render multiple times. Keep tabs on the following issue:
+    //   https://github.com/joepuzzo/informed/issues/138
+    // If they resolve it or we move away from informed we can probably get some
+    // extra performance.
+    const formState = useFormState();
+
+    const billingAddressFields = !formState.values.addresses_same ? (
+        <Fragment>
+            <div className={classes.street0}>
+                <Field label="Street">
+                    <TextInput
+                        id={classes.street0}
+                        field="street[0]"
+                        validate={isRequired}
+                    />
+                </Field>
+            </div>
+            <div className={classes.city}>
+                <Field label="City">
+                    <TextInput
+                        id={classes.city}
+                        field="city"
+                        validate={isRequired}
+                    />
+                </Field>
+            </div>
+            <div className={classes.region_code}>
+                <Field label="State">
+                    <TextInput
+                        id={classes.region_code}
+                        field="region_code"
+                        validate={combine([
+                            isRequired,
+                            [hasLengthExactly, 2],
+                            [validateRegionCode, countries]
+                        ])}
+                    />
+                </Field>
+            </div>
+            <div className={classes.postcode}>
+                <Field label="ZIP">
+                    <TextInput
+                        id={classes.postcode}
+                        field="postcode"
+                        validate={isRequired}
+                    />
+                </Field>
+            </div>
+        </Fragment>
+    ) : null;
+
+    const handleError = useCallback(() => {
+        setIsSubmitting(false);
+    }, [setIsSubmitting]);
+
+    // The success callback. Unfortunately since form state is created first and
+    // then modified when using initialValues any component who uses this
+    // callback will be rendered multiple times on first render. See above
+    // comments for more info.
+    const handleSuccess = useCallback(
+        value => {
+            setIsSubmitting(false);
+            const sameAsShippingAddress = formState.values['addresses_same'];
+            let billingAddress;
+            if (!sameAsShippingAddress) {
+                billingAddress = {
+                    city: formState.values['city'],
+                    postcode: formState.values['postcode'],
+                    region_code: formState.values['region_code'],
+                    street: formState.values['street']
+                };
+            } else {
+                billingAddress = {
+                    sameAsShippingAddress
+                };
+            }
+            submit({
+                billingAddress,
+                paymentMethod: {
+                    code: 'braintree',
+                    data: value
+                }
+            });
+        },
+        [formState.values, setIsSubmitting, submit]
+    );
+
+    return (
+        <Fragment>
+            <div className={classes.body}>
+                <h2 className={classes.heading}>Billing Information</h2>
+                <div className={classes.braintree}>
+                    <BraintreeDropin
+                        shouldRequestPaymentNonce={isSubmitting}
+                        onError={handleError}
+                        onSuccess={handleSuccess}
+                    />
+                </div>
+                <div className={classes.address_check}>
+                    <Checkbox
+                        field="addresses_same"
+                        label="Billing address same as shipping address"
+                    />
+                </div>
+                {billingAddressFields}
+            </div>
+            <div className={classes.footer}>
+                <Button className={classes.button} onClick={cancel}>
+                    Cancel
+                </Button>
+                <Button
+                    className={classes.button}
+                    priority="high"
+                    type="submit"
+                    disabled={submitting}
+                >
+                    Use Card
+                </Button>
+            </div>
+        </Fragment>
+    );
+};
+
+PaymentsFormItems.propTypes = {
+    cancel: func.isRequired,
+    classes: shape({
+        address_check: string,
+        body: string,
+        button: string,
+        braintree: string,
+        city: string,
+        footer: string,
+        heading: string,
+        postcode: string,
+        region_code: string,
+        street0: string
+    }),
+    countries: array,
+    isSubmitting: bool,
+    setIsSubmitting: func.isRequired,
+    submit: func.isRequired,
+    submitting: bool
+};
+
+export default PaymentsFormItems;
